Bug #2157165 - [A13解耦项目][K515][sharkle]camera bringup
[root cause  ]:A13解耦分支编译失败
[changes     ]:修复A13解耦分支编译失败
[side effects]:none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:9832e
[test case]:camera
[reviewers]:xiaoming.ren,yulong.fan

commit_template_version:v1
[change_type ] feature_bugfix
[tag_product ] common

+++ core/dcam_if_r4p0_isp_r6p11/src/dcam_core.c
@@ -1932,7 +1932,7 @@ static void sprd_img_set_vcm_pos(struct sprd_img_vcm_param *vcm_info)
 	struct sensor_i2c_tag i2c_tab;
 
 	pr_debug("i2c set vcm pos %d\n", vcm_info->next_vcm_pos);
-	i2c_tab.i2c_data = vcm_info->vcm_i2c_data;
+	i2c_tab.i2c_data = (uint64_t)vcm_info->vcm_i2c_data;
 	i2c_tab.i2c_count = vcm_info->vcm_i2c_count;
 	i2c_tab.slave_addr = vcm_info->vcm_slave_addr;
